docs/conf.py: Ignore some missing references in nitpick mode
We enable Sphinx's 'nitpick' mode via the command line -n switch, which
means it warns about references to things it doesn't know about. Add
a whitelist of expected-not-to-be-present types, to avoid false
positives when C function prototypes use standard integer types.

If this whitelist gets too long and unmanageable we might be better
off switching to non-nitpick mode instead.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>